refactor: widgets Card 컴포넌트 이름 충돌 해결
- post/Card.tsx: Card → PostCard
- thread/Card.tsx: Card → ThreadCard
- 모든 index.ts export 수정
- List.tsx에서 PostCard import 수정

shared/ui/card의 Card와 이름 충돌 방지
