+++ src/entities/thread/index.ts
@@ -1,2 +1,1 @@
-export { ThreadCard } from "./ui/ThreadCard";
 export { getThreads, type Thread } from "./lib/threads";
