+++ src/entities/user/index.ts
@@ -1,4 +1,3 @@
 export type { User, UserBio, UserSocial } from './model/types';
-export { Bio } from './ui/Bio';
 export { UserAvatar } from './ui/UserAvatar';
 export { getUserAvatarUrl, getUserDisplayName, getUserBioDescription, validateSocialLinks } from './lib/utils';
