+++ src/widgets/layout/index.ts
@@ -1,1 +1,1 @@
-export { Layout } from './ui/Layout';
+export { Layout } from './Layout';
