--- src/widgets/page-footer/index.ts
@@ -1,1 +1,0 @@
-export { PageFooter } from './ui/PageFooter';
