--- src/widgets/page-header/index.ts
@@ -1,1 +1,0 @@
-export { PageHeader } from './ui/PageHeader';
