+++ src/widgets/page/index.ts
@@ -1,0 +1,2 @@
+export { Header } from './Header';
+export { Footer } from './Footer';
