--- src/widgets/post-card/index.ts
@@ -1,1 +1,0 @@
-export { PostCard } from './ui/PostCard';
