--- src/widgets/post-content/index.ts
@@ -1,1 +1,0 @@
-export { PostContent } from './ui/PostContent';
