--- src/widgets/post-list/index.ts
@@ -1,1 +1,0 @@
-export { PostList } from './ui/PostList';
