+++ src/widgets/post/index.ts
@@ -1,0 +1,3 @@
+export { PostCard } from './Card';
+export { List } from './List';
+export { Content } from './Content';
