+++ src/widgets/thread/ThreadCard.tsx
