+++ src/widgets/thread/index.ts
@@ -1,0 +1,1 @@
+export { ThreadCard } from "./ThreadCard";
