+++ src/widgets/user/index.ts
@@ -1,0 +1,1 @@
+export { Bio } from './Bio';
